Fix filterDump failure on base64 encoded image
The content of a css url directive could be a base64 encoded image. The
length of the string can be long. ( >4096). On some plateforms the
filename length is limited to less than 4096. In this case the fonction
filterDump fails because of using `file_exists` (in checkPath) on string
longer than 4096.

To prevent this, you need to check if the url directive begins with
"data:"